requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
json-repair
